refactor(get_describecluster_info): move from tester to cluster
Method's get_describecluster_info() functionality places the cluster
rather than the tester. By doing this, I am able to use it from node/cluster
rather than `ClusterTester`
